@@ -1,16 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Net;
+using System.Text;
 using System.Threading.Tasks;
 using Fido2NetLib;
 using Fido2NetLib.Development;
 using Fido2NetLib.Objects;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Caching.Distributed;
-using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Options;
 
 namespace Fido2Demo
 {
@@ -27,26 +25,28 @@ namespace Fido2Demo
          */
         private static readonly DevelopmentInMemoryStore DemoStorage = new DevelopmentInMemoryStore();
 
-        private IFido2 _fido2;
-        private string _origin;
+        private readonly IFido2 _fido2;
+        private readonly string _origin;
 
-        public TestController(IConfiguration config)
+        public TestController(IOptions<Fido2Configuration> fido2Configuration)
         {
-            _origin = config["fido2:origin"];
+            _origin = fido2Configuration.Value.Origin;
 
-            _fido2 = new Fido2(new Fido2Configuration()
+            _fido2 = new Fido2(new Fido2Configuration
             {
-                ServerDomain = config["fido2:serverDomain"],
-                ServerName = "Fido2 test",
+                ServerDomain = fido2Configuration.Value.ServerDomain,
+                ServerName = fido2Configuration.Value.ServerName,
                 Origin = _origin
-            }, ConformanceTesting.MetadataServiceInstance(config["fido2:MDSCacheDirPath"] + @"\Conformance", _origin));
+            }, 
+            ConformanceTesting.MetadataServiceInstance(
+                System.IO.Path.Combine(fido2Configuration.Value.MDSCacheDirPath, @"Conformance"), _origin)
+            );
         }
 
         [HttpPost]
         [Route("/attestation/options")]
         public JsonResult MakeCredentialOptionsTest([FromBody] TEST_MakeCredentialParams opts)
         {
-
             var attType = opts.Attestation;
 
             var username = new byte[] { };
@@ -55,9 +55,9 @@ namespace Fido2Demo
             {
                 username = Base64Url.Decode(opts.Username);
             }
-            catch(FormatException)
+            catch (FormatException)
             {
-                username = System.Text.Encoding.UTF8.GetBytes(opts.Username);
+                username = Encoding.UTF8.GetBytes(opts.Username);
             }
 
             // 1. Get user from DB by username (in our example, auto create missing users)
@@ -72,11 +72,10 @@ namespace Fido2Demo
             var existingKeys = DemoStorage.GetCredentialsByUser(user).Select(c => c.Descriptor).ToList();
 
             //var exts = new AuthenticationExtensionsClientInputs() { Extensions = true, UserVerificationIndex = true, Location = true, UserVerificationMethod = true, BiometricAuthenticatorPerformanceBounds = new AuthenticatorBiometricPerfBounds { FAR = float.MaxValue, FRR = float.MaxValue } };
-            var exts = new AuthenticationExtensionsClientInputs() {  };
-            if (null != opts.Extensions
-                && null != opts.Extensions.Example)
-
+            var exts = new AuthenticationExtensionsClientInputs() { };
+            if (null != opts.Extensions && null != opts.Extensions.Example)
                 exts.Example = opts.Extensions.Example;
+
             // 3. Create options
             var options = _fido2.RequestNewCredential(user, existingKeys, opts.AuthenticatorSelection, opts.Attestation, exts);
 
@@ -100,9 +99,7 @@ namespace Fido2Demo
             IsCredentialIdUniqueToUserAsyncDelegate callback = async (IsCredentialIdUniqueToUserParams args) =>
             {
                 var users = await DemoStorage.GetUsersByCredentialIdAsync(args.CredentialId);
-                if (users.Count > 0) return false;
-
-                return true;
+                return users.Count <= 0;
             };
 
             // 2. Verify and make the credentials
@@ -128,18 +125,30 @@ namespace Fido2Demo
             var username = assertionClientParams.Username;
             // 1. Get user from DB
             var user = DemoStorage.GetUser(username);
-            if (user == null) return NotFound("username was not registered");
+            if (user == null)
+                return NotFound("username was not registered");
 
             // 2. Get registered credentials from database
             var existingCredentials = DemoStorage.GetCredentialsByUser(user).Select(c => c.Descriptor).ToList();
 
             var uv = assertionClientParams.UserVerification;
-            if (null != assertionClientParams.authenticatorSelection) uv = assertionClientParams.authenticatorSelection.UserVerification;
+            if (null != assertionClientParams.authenticatorSelection)
+                uv = assertionClientParams.authenticatorSelection.UserVerification;
 
-            var exts = new AuthenticationExtensionsClientInputs() { AppID = _origin, SimpleTransactionAuthorization = "FIDO", GenericTransactionAuthorization = new TxAuthGenericArg { ContentType = "text/plain", Content = new byte[] { 0x46, 0x49, 0x44, 0x4F } }, UserVerificationIndex = true, Location = true, UserVerificationMethod = true };
-            if (null != assertionClientParams.Extensions
-                && null != assertionClientParams.Extensions.Example)
-
+            var exts = new AuthenticationExtensionsClientInputs
+            { 
+                AppID = _origin,
+                SimpleTransactionAuthorization = "FIDO",
+                GenericTransactionAuthorization = new TxAuthGenericArg 
+                { 
+                    ContentType = "text/plain",
+                    Content = new byte[] { 0x46, 0x49, 0x44, 0x4F }
+                }, 
+                UserVerificationIndex = true, 
+                Location = true, 
+                UserVerificationMethod = true
+            };
+            if (null != assertionClientParams.Extensions && null != assertionClientParams.Extensions.Example)
                 exts.Example = assertionClientParams.Extensions.Example;
 
             // 3. Create options
@@ -191,11 +200,6 @@ namespace Fido2Demo
 
             // 7. return OK to client
             return Json(testRes);
-        }
-
-        private byte[] GetTokenBindingId()
-        {
-            return Request.HttpContext.Features.Get<ITlsTokenBindingFeature>()?.GetProvidedTokenBindingId();
         }
 
         /// <summary>
