@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using System.Text;
 using Newtonsoft.Json;
 
@@ -9,27 +10,24 @@ namespace Fido2NetLib
     /// </summary>
     public class AuthenticatorResponse
     {
-
         protected AuthenticatorResponse(byte[] clientDataJson)
         {
-            if (null == clientDataJson) throw new Fido2VerificationException("clientDataJson cannot be null");
+            if (null == clientDataJson)
+                throw new Fido2VerificationException("clientDataJson cannot be null");
             var stringx = Encoding.UTF8.GetString(clientDataJson);
 
-            AuthenticatorResponse response = null;
+            AuthenticatorResponse response;
             try
             {
-                response = Newtonsoft.Json.JsonConvert.DeserializeObject<AuthenticatorResponse>(stringx);
+                response = JsonConvert.DeserializeObject<AuthenticatorResponse>(stringx);
             }
-            catch (System.Exception e)//Newtonsoft.Json.JsonReaderException) 
+            catch (Exception e) when (e is JsonReaderException || e is JsonSerializationException)
             {
-                if (e is JsonReaderException || e is JsonSerializationException)
-                {
-                    throw new Fido2VerificationException("Malformed clientDataJson");
-                }
-                else throw;
+                throw new Fido2VerificationException("Malformed clientDataJson");
             }
 
-            if (null == response) throw new Fido2VerificationException("Deserialized authenticator response cannot be null");
+            if (null == response)
+                throw new Fido2VerificationException("Deserialized authenticator response cannot be null");
             Type = response.Type;
             Challenge = response.Challenge;
             Origin = response.Origin;
@@ -55,13 +53,18 @@ namespace Fido2NetLib
 
         protected void BaseVerify(string expectedOrigin, byte[] originalChallenge, byte[] requestTokenBindingId)
         {
-            if (null == Challenge) throw new Fido2VerificationException("Challenge cannot be null");
+            if (null == Challenge)
+                throw new Fido2VerificationException("Challenge cannot be null");
+
             // verify challenge is same
-            if (!Challenge.SequenceEqual(originalChallenge)) throw new Fido2VerificationException("Challenge not equal to original challenge");
+            if (!Challenge.SequenceEqual(originalChallenge))
+                throw new Fido2VerificationException("Challenge not equal to original challenge");
 
-            if (Origin != expectedOrigin) throw new Fido2VerificationException("Origin not equal to original origin");
+            if (Origin != expectedOrigin)
+                throw new Fido2VerificationException($"Origin {Origin} not equal to original origin {expectedOrigin}");
 
-            if (Type != "webauthn.create" && Type != "webauthn.get") throw new Fido2VerificationException($"Type not equal to 'webauthn.create' or 'webauthn.get'. Was: '{Type}'");
+            if (Type != "webauthn.create" && Type != "webauthn.get")
+                throw new Fido2VerificationException($"Type not equal to 'webauthn.create' or 'webauthn.get'. Was: '{Type}'");
 
             if (TokenBinding != null)
             {
