@@ -30,7 +30,10 @@ namespace Fido2NetLib
         /// </summary>
         /// <returns></returns>
         /// <param name="excludeCredentials">Recommended. This member is intended for use by Relying Parties that wish to limit the creation of multiple credentials for the same account on a single authenticator.The client is requested to return an error if the new credential would be created on an authenticator that also contains one of the credentials enumerated in this parameter.</param>
-        public CredentialCreateOptions RequestNewCredential(Fido2User user, List<PublicKeyCredentialDescriptor> excludeCredentials, AuthenticationExtensionsClientInputs extensions = null)
+        public CredentialCreateOptions RequestNewCredential(
+            Fido2User user,
+            List<PublicKeyCredentialDescriptor> excludeCredentials,
+            AuthenticationExtensionsClientInputs extensions = null)
         {
             return RequestNewCredential(user, excludeCredentials, AuthenticatorSelection.Default, AttestationConveyancePreference.None, extensions);
         }
@@ -41,7 +44,12 @@ namespace Fido2NetLib
         /// <returns></returns>
         /// <param name="attestationPreference">This member is intended for use by Relying Parties that wish to express their preference for attestation conveyance. The default is none.</param>
         /// <param name="excludeCredentials">Recommended. This member is intended for use by Relying Parties that wish to limit the creation of multiple credentials for the same account on a single authenticator.The client is requested to return an error if the new credential would be created on an authenticator that also contains one of the credentials enumerated in this parameter.</param>
-        public CredentialCreateOptions RequestNewCredential(Fido2User user, List<PublicKeyCredentialDescriptor> excludeCredentials, AuthenticatorSelection authenticatorSelection, AttestationConveyancePreference attestationPreference, AuthenticationExtensionsClientInputs extensions = null)
+        public CredentialCreateOptions RequestNewCredential(
+            Fido2User user,
+            List<PublicKeyCredentialDescriptor> excludeCredentials,
+            AuthenticatorSelection authenticatorSelection,
+            AttestationConveyancePreference attestationPreference,
+            AuthenticationExtensionsClientInputs extensions = null)
         {
             // note: I have no idea if this crypto is ok...
             var challenge = new byte[_config.ChallengeSize];
@@ -57,20 +65,32 @@ namespace Fido2NetLib
         /// <param name="attestationResponse"></param>
         /// <param name="origChallenge"></param>
         /// <returns></returns>
-        public async Task<CredentialMakeResult> MakeNewCredentialAsync(AuthenticatorAttestationRawResponse attestationResponse, CredentialCreateOptions origChallenge, IsCredentialIdUniqueToUserAsyncDelegate isCredentialIdUniqueToUser, byte[] requestTokenBindingId = null)
+        public async Task<CredentialMakeResult> MakeNewCredentialAsync(
+            AuthenticatorAttestationRawResponse attestationResponse,
+            CredentialCreateOptions origChallenge,
+            IsCredentialIdUniqueToUserAsyncDelegate isCredentialIdUniqueToUser,
+            byte[] requestTokenBindingId = null)
         {
             var parsedResponse = AuthenticatorAttestationResponse.Parse(attestationResponse);
             var success = await parsedResponse.VerifyAsync(origChallenge, _config, isCredentialIdUniqueToUser, _metadataService, requestTokenBindingId);
 
             // todo: Set Errormessage etc.
-            return new CredentialMakeResult { Status = "ok", ErrorMessage = string.Empty, Result = success };
+            return new CredentialMakeResult 
+            { 
+                Status = "ok", 
+                ErrorMessage = string.Empty, 
+                Result = success 
+            };
         }
 
         /// <summary>
         /// Returns AssertionOptions including a challenge to the browser/authr to assert existing credentials and authenticate a user.
         /// </summary>
         /// <returns></returns>
-        public AssertionOptions GetAssertionOptions(IEnumerable<PublicKeyCredentialDescriptor> allowedCredentials, UserVerificationRequirement? userVerification, AuthenticationExtensionsClientInputs extensions = null)
+        public AssertionOptions GetAssertionOptions(
+            IEnumerable<PublicKeyCredentialDescriptor> allowedCredentials,
+            UserVerificationRequirement? userVerification,
+            AuthenticationExtensionsClientInputs extensions = null)
         {
             var challenge = new byte[_config.ChallengeSize];
             _crypto.GetBytes(challenge);
@@ -83,11 +103,22 @@ namespace Fido2NetLib
         /// Verifies the assertion response from the browser/authr to assert existing credentials and authenticate a user.
         /// </summary>
         /// <returns></returns>
-        public async Task<AssertionVerificationResult> MakeAssertionAsync(AuthenticatorAssertionRawResponse assertionResponse, AssertionOptions originalOptions, byte[] storedPublicKey, uint storedSignatureCounter, IsUserHandleOwnerOfCredentialIdAsync isUserHandleOwnerOfCredentialIdCallback, byte[] requestTokenBindingId = null)
+        public async Task<AssertionVerificationResult> MakeAssertionAsync(
+            AuthenticatorAssertionRawResponse assertionResponse,
+            AssertionOptions originalOptions,
+            byte[] storedPublicKey,
+            uint storedSignatureCounter,
+            IsUserHandleOwnerOfCredentialIdAsync isUserHandleOwnerOfCredentialIdCallback,
+            byte[] requestTokenBindingId = null)
         {
             var parsedResponse = AuthenticatorAssertionResponse.Parse(assertionResponse);
 
-            var result = await parsedResponse.VerifyAsync(originalOptions, _config.Origin, storedPublicKey, storedSignatureCounter, isUserHandleOwnerOfCredentialIdCallback, requestTokenBindingId);
+            var result = await parsedResponse.VerifyAsync(originalOptions,
+                                                          _config.Origin,
+                                                          storedPublicKey,
+                                                          storedSignatureCounter,
+                                                          isUserHandleOwnerOfCredentialIdCallback,
+                                                          requestTokenBindingId);
 
             return result;
         }
