@@ -1,19 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Net.Http;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Tasks;
-using Fido2NetLib;
 using Microsoft.IdentityModel.Tokens;
 using Newtonsoft.Json;
-//using static Fido2NetLib.MDSMetadata;
+using Newtonsoft.Json.Linq;
 
 namespace Fido2NetLib
 {
-    class MDSGetEndpointResponse
+    internal class MDSGetEndpointResponse
     {
         [JsonProperty("status", Required = Required.Always)]
         public string Status { get; set; }
@@ -67,9 +67,12 @@ namespace Fido2NetLib
             var tocAlg = await GetTocAlg();
 
             var statementBytes = Base64Url.Decode(statementBase64Url);
-            var statementString = System.Text.Encoding.UTF8.GetString(statementBytes, 0, statementBytes.Length);
+            var statementString = Encoding.UTF8.GetString(statementBytes, 0, statementBytes.Length);
             var statement = JsonConvert.DeserializeObject<MetadataStatement>(statementString);
-            statement.Hash = Base64Url.Encode(CryptoUtils.GetHasher(new HashAlgorithmName(tocAlg)).ComputeHash(System.Text.Encoding.UTF8.GetBytes(statementBase64Url)));
+            using(HashAlgorithm hasher = CryptoUtils.GetHasher(new HashAlgorithmName(tocAlg)))
+            {
+                statement.Hash = Base64Url.Encode(hasher.ComputeHash(Encoding.UTF8.GetBytes(statementBase64Url)));
+            }
 
             return statement;
         }
@@ -81,14 +84,16 @@ namespace Fido2NetLib
                 endpoint = _origin
             };
 
-            var content = new StringContent(JsonConvert.SerializeObject(req), System.Text.Encoding.UTF8, "application/json");
+            var content = new StringContent(JsonConvert.SerializeObject(req), Encoding.UTF8, "application/json");
             var response = await _httpClient.PostAsync(_getEndpointsUrl, content);
             var result = JsonConvert.DeserializeObject<MDSGetEndpointResponse>(await response.Content.ReadAsStringAsync());
-            var conformanceEndpoints = new System.Collections.Generic.List<string>(result.Result);
+            var conformanceEndpoints = new List<string>(result.Result);
 
-            var combinedToc = new MetadataTOCPayload();
-            combinedToc.Number = -1;
-            combinedToc.NextUpdate = "2099-08-07";
+            var combinedToc = new MetadataTOCPayload
+            {
+                Number = -1,
+                NextUpdate = "2099-08-07"
+            };
 
             var entries = new List<MetadataTOCPayloadEntry>();
 
@@ -107,8 +112,10 @@ namespace Fido2NetLib
                     continue;
                 }
                 
-                if(string.Compare(toc.NextUpdate, combinedToc.NextUpdate) < 0) combinedToc.NextUpdate = toc.NextUpdate;
-                if (combinedToc.Number < toc.Number) combinedToc.Number = toc.Number;
+                if(string.Compare(toc.NextUpdate, combinedToc.NextUpdate) < 0)
+                    combinedToc.NextUpdate = toc.NextUpdate;
+                if (combinedToc.Number < toc.Number)
+                    combinedToc.Number = toc.Number;
 
                 foreach (var entry in toc.Entries)
                 {
@@ -132,13 +139,12 @@ namespace Fido2NetLib
 
         public async Task<MetadataTOCPayload> DeserializeAndValidateToc(string toc)
         {
-            var jwtToken = new System.IdentityModel.Tokens.Jwt.JwtSecurityToken(toc);
+            var jwtToken = new JwtSecurityToken(toc);
             _tocAlg = jwtToken.Header["alg"] as string;
 
             var keys = new List<SecurityKey>();
 
-            var keyStrings = (jwtToken.Header["x5c"] as Newtonsoft.Json.Linq.JArray)
-                .Values<string>();
+            var keyStrings = (jwtToken.Header["x5c"] as JArray).Values<string>();
 
             foreach(var keyString in keyStrings)
             {
@@ -171,18 +177,18 @@ namespace Fido2NetLib
                 IssuerSigningKeys = keys,
             };
 
-            var tokenHandler = new System.IdentityModel.Tokens.Jwt.JwtSecurityTokenHandler();
+            var tokenHandler = new JwtSecurityTokenHandler();
 
             tokenHandler.ValidateToken(
                 toc,
                 validationParameters,
                 out var validatedToken);
 
-            var payload = ((System.IdentityModel.Tokens.Jwt.JwtSecurityToken)validatedToken).Payload.SerializeToJson();
+            var payload = ((JwtSecurityToken)validatedToken).Payload.SerializeToJson();
 
-            chain.ChainPolicy.ExtraStore.Add(new X509Certificate2(System.Convert.FromBase64String((jwtToken.Header["x5c"] as Newtonsoft.Json.Linq.JArray).Values<string>().Last())));
+            chain.ChainPolicy.ExtraStore.Add(new X509Certificate2(Convert.FromBase64String((jwtToken.Header["x5c"] as JArray).Values<string>().Last())));
 
-            var valid = chain.Build(new X509Certificate2(System.Convert.FromBase64String((jwtToken.Header["x5c"] as Newtonsoft.Json.Linq.JArray).Values<string>().First())));
+            var valid = chain.Build(new X509Certificate2(Convert.FromBase64String((jwtToken.Header["x5c"] as JArray).Values<string>().First())));
             // if the root is trusted in the context we are running in, valid should be true here
             if (!valid)
             {
@@ -192,14 +198,15 @@ namespace Fido2NetLib
                     {
                         var cdp = CryptoUtils.CDPFromCertificateExts(element.Certificate.Extensions);
                         var crlFile = await DownloadDataAsync(cdp);
-                        if (true == CryptoUtils.IsCertInCRL(crlFile, element.Certificate)) throw new Fido2VerificationException(string.Format("Cert {0} found in CRL {1}", element.Certificate.Subject, cdp));
+                        if (true == CryptoUtils.IsCertInCRL(crlFile, element.Certificate))
+                            throw new Fido2VerificationException(string.Format("Cert {0} found in CRL {1}", element.Certificate.Subject, cdp));
                     }
                 }
 
                 // otherwise we have to manually validate that the root in the chain we are testing is the root we downloaded
                 if (root.Thumbprint == chain.ChainElements[chain.ChainElements.Count - 1].Certificate.Thumbprint &&
                     // and that the number of elements in the chain accounts for what was in x5c plus the root we added
-                    chain.ChainElements.Count == ((jwtToken.Header["x5c"] as Newtonsoft.Json.Linq.JArray).Count + 1) &&
+                    chain.ChainElements.Count == ((jwtToken.Header["x5c"] as JArray).Count + 1) &&
                     // and that the root cert has exactly one status listed against it
                     chain.ChainElements[chain.ChainElements.Count - 1].ChainElementStatus.Length == 1 &&
                     // and that that status is a status of exactly UntrustedRoot
