@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
-using System.Text;
 using System.Threading.Tasks;
 using Fido2NetLib.AttestationFormat;
 using Newtonsoft.Json;
@@ -45,12 +44,13 @@ namespace Fido2NetLib
 
         public async Task<MetadataStatement> GetMetadataStatement(MetadataTOCPayloadEntry entry)
         {
-            if (_toc == null) await GetToc();
+            if (_toc == null)
+                await GetToc();
 
-            Guid parsedAaGuid;
-            if (!string.IsNullOrEmpty(entry.AaGuid) && Guid.TryParse(entry.AaGuid, out parsedAaGuid))
+            if (!string.IsNullOrEmpty(entry.AaGuid) && Guid.TryParse(entry.AaGuid, out Guid parsedAaGuid))
             {
-                if (_entries.ContainsKey(parsedAaGuid)) return _entries[parsedAaGuid].MetadataStatement;
+                if (_entries.ContainsKey(parsedAaGuid))
+                    return _entries[parsedAaGuid].MetadataStatement;
             }
 
             return null;
@@ -68,8 +68,23 @@ namespace Fido2NetLib
             {
                 AaGuid = "f8a011f3-8c0a-4d15-8006-17111f9edc7d",
                 Hash = "",
-                StatusReports = new StatusReport[] { new StatusReport() { Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED } },
-                MetadataStatement = new MetadataStatement() { AttestationTypes = new ushort[] { (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL }, Hash = "", Description = "Yubico YubiKey FIDO2", AttestationRootCertificates = new string[] { YUBICO_ROOT } }
+                StatusReports = new StatusReport[]
+                {
+                    new StatusReport() { Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED }
+                },
+                MetadataStatement = new MetadataStatement
+                {
+                    AttestationTypes = new ushort[]
+                    {
+                        (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL
+                    },
+                    Hash = "",
+                    Description = "Yubico YubiKey FIDO2",
+                    AttestationRootCertificates = new string[]
+                    {
+                        YUBICO_ROOT
+                    }
+                }
             };
             _entries.Add(new Guid(yubico.AaGuid), yubico);
 
@@ -78,8 +93,26 @@ namespace Fido2NetLib
             {
                 AaGuid = "cb69481e-8ff7-4039-93ec-0a2729a154a8",
                 Hash = "",
-                StatusReports = new StatusReport[] { new StatusReport() { Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED } },
-                MetadataStatement = new MetadataStatement() { AttestationTypes = new ushort[] { (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL }, Hash = "", Description = "Yubico YubiKey 5 USB", AttestationRootCertificates = new string[] { YUBICO_ROOT } }
+                StatusReports = new StatusReport[]
+                {
+                    new StatusReport
+                    {
+                        Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED
+                    }
+                },
+                MetadataStatement = new MetadataStatement
+                {
+                    AttestationTypes = new ushort[]
+                    {
+                        (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL
+                    },
+                    Hash = "",
+                    Description = "Yubico YubiKey 5 USB",
+                    AttestationRootCertificates = new string[]
+                    {
+                        YUBICO_ROOT
+                    }
+                }
             };
             _entries.Add(new Guid(yubikey5usb.AaGuid), yubikey5usb);
 
@@ -87,8 +120,26 @@ namespace Fido2NetLib
             {
                 AaGuid = "fa2b99dc-9e39-4257-8f92-4a30d23c4118",
                 Hash = "",
-                StatusReports = new StatusReport[] { new StatusReport() { Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED } },
-                MetadataStatement = new MetadataStatement() { AttestationTypes = new ushort[] { (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL }, Hash = "", Description = "Yubico YubiKey 5 NFC", AttestationRootCertificates = new string[] { YUBICO_ROOT } }
+                StatusReports = new StatusReport[]
+                {
+                    new StatusReport
+                    {
+                        Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED
+                    }
+                },
+                MetadataStatement = new MetadataStatement
+                {
+                    AttestationTypes = new ushort[]
+                    {
+                        (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL
+                    },
+                    Hash = "",
+                    Description = "Yubico YubiKey 5 NFC",
+                    AttestationRootCertificates = new string[]
+                    {
+                        YUBICO_ROOT
+                    }
+                }
             };
             _entries.Add(new Guid(yubikey5nfc.AaGuid), yubikey5nfc);
 
@@ -97,13 +148,15 @@ namespace Fido2NetLib
                 AaGuid = "6d44ba9b-f6ec-2e49-b930-0c8fe920cb73",
                 Hash = "",
                 StatusReports = new StatusReport[] { new StatusReport() { Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED } },
-                MetadataStatement = new MetadataStatement() {
+                MetadataStatement = new MetadataStatement
+                {
                     Description = "Yubico Security Key NFC",
                     Icon = "data:image/svg+xml;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz4KPCEtLSBHZW5lcmF0b3I6IEFkb2JlIElsbHVzdHJhdG9yIDIzLjAuMSwgU1ZHIEV4cG9ydCBQbHVnLUluIC4gU1ZHIFZlcnNpb246IDYuMDAgQnVpbGQgMCkgIC0tPgo8c3ZnIHZlcnNpb249IjEuMSIgaWQ9Ill1YmljbyIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIiB4bWxuczp4bGluaz0iaHR0cDovL3d3dy53My5vcmcvMTk5OS94bGluayIgeD0iMHB4IiB5PSIwcHgiCgkgdmlld0JveD0iMCAwIDc2OCA3NjgiIHN0eWxlPSJlbmFibGUtYmFja2dyb3VuZDpuZXcgMCAwIDc2OCA3Njg7IiB4bWw6c3BhY2U9InByZXNlcnZlIj4KPHN0eWxlIHR5cGU9InRleHQvY3NzIj4KCS5zdDB7ZmlsbDojOUFDQTNDO30KPC9zdHlsZT4KPHBvbHlnb24gaWQ9IlkiIGNsYXNzPSJzdDAiIHBvaW50cz0iMjE4LjQzLDIxMS44MSAzMTYuNDksMjExLjgxIDM4Ni41Miw0MDAuMDcgNDUzLjIsMjExLjgxIDU0OS41NywyMTEuODEgMzg3LjA4LDYxMS44NiAKCTI4Ni4yMyw2MTEuODYgMzMyLjE3LDUwMi4wNCAiLz4KPHBhdGggaWQ9IkNpcmNsZV8xXyIgY2xhc3M9InN0MCIgZD0iTTM4NCwwQzE3MS45MiwwLDAsMTcxLjkyLDAsMzg0czE3MS45MiwzODQsMzg0LDM4NHMzODQtMTcxLjkyLDM4NC0zODRTNTk2LjA4LDAsMzg0LDB6CgkgTTM4NCw2OTMuNThDMjEzLjAyLDY5My41OCw3NC40Miw1NTQuOTgsNzQuNDIsMzg0UzIxMy4wMiw3NC40MiwzODQsNzQuNDJTNjkzLjU4LDIxMy4wMiw2OTMuNTgsMzg0UzU1NC45OCw2OTMuNTgsMzg0LDY5My41OHoiLz4KPC9zdmc+Cg==",
                     AttachmentHint = 6,
                     AttestationTypes = new ushort[] { (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL },
                     Hash = "",
-                    AttestationRootCertificates = new string[] {
+                    AttestationRootCertificates = new string[]
+                    {
                         YUBICO_ROOT
                     }
                 }
@@ -114,32 +167,88 @@ namespace Fido2NetLib
             {
                 AaGuid = "6028B017-B1D4-4C02-B4B3-AFCDAFC96BB2",
                 Hash = "",
-                StatusReports = new StatusReport[] { new StatusReport() { Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED } },
-                MetadataStatement = new MetadataStatement() { AttestationTypes = new ushort[] { (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL }, Hash = "", Description = "Windows Hello software authenticator" }
+                StatusReports = new StatusReport[]
+                {
+                    new StatusReport
+                    {
+                        Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED
+                    }
+                },
+                MetadataStatement = new MetadataStatement
+                {
+                    AttestationTypes = new ushort[]
+                    {
+                        (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL
+                    },
+                    Hash = "",
+                    Description = "Windows Hello software authenticator"
+                }
             };
             _entries.Add(new Guid(msftWhfbSoftware.AaGuid), msftWhfbSoftware);
             var msftWhfbSoftwareVbs = new MetadataTOCPayloadEntry
             {
                 AaGuid = "6E96969E-A5CF-4AAD-9B56-305FE6C82795",
                 Hash = "",
-                StatusReports = new StatusReport[] { new StatusReport() { Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED } },
-                MetadataStatement = new MetadataStatement() { AttestationTypes = new ushort[] { (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL }, Hash = "", Description = "Windows Hello VBS software authenticator" }
+                StatusReports = new StatusReport[]
+                {
+                    new StatusReport
+                    {
+                        Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED
+                    }
+                },
+                MetadataStatement = new MetadataStatement
+                {
+                    AttestationTypes = new ushort[]
+                    {
+                        (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL
+                    },
+                    Hash = "",
+                    Description = "Windows Hello VBS software authenticator"
+                }
             };
             _entries.Add(new Guid(msftWhfbSoftwareVbs.AaGuid), msftWhfbSoftwareVbs);
             var msftWhfbHardware = new MetadataTOCPayloadEntry
             {
                 AaGuid = "08987058-CADC-4B81-B6E1-30DE50DCBE96",
                 Hash = "",
-                StatusReports = new StatusReport[] { new StatusReport() { Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED } },
-                MetadataStatement = new MetadataStatement() { AttestationTypes = new ushort[] { (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL }, Hash = "", Description = "Windows Hello hardware authenticator" }
+                StatusReports = new StatusReport[]
+                {
+                    new StatusReport
+                    {
+                        Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED
+                    }
+                },
+                MetadataStatement = new MetadataStatement
+                {
+                    AttestationTypes = new ushort[]
+                    {
+                        (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL
+                    },
+                    Hash = "",
+                    Description = "Windows Hello hardware authenticator"
+                }
             };
             _entries.Add(new Guid(msftWhfbHardware.AaGuid), msftWhfbHardware);
             var msftWhfbHardwareVbs = new MetadataTOCPayloadEntry
             {
                 AaGuid = "9DDD1817-AF5A-4672-A2B9-3E3DD95000A9",
                 Hash = "",
-                StatusReports = new StatusReport[] { new StatusReport() { Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED } },
-                MetadataStatement = new MetadataStatement() { AttestationTypes = new ushort[] { (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL }, Hash = "", Description = "Windows Hello VBS hardware authenticator" }
+                StatusReports = new StatusReport[]
+                {
+                    new StatusReport
+                    {
+                        Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED
+                    }
+                },
+                MetadataStatement = new MetadataStatement
+                {
+                    AttestationTypes = new ushort[]
+                    {
+                        (ushort)MetadataAttestationType.ATTESTATION_BASIC_FULL
+                    },
+                    Hash = "",
+                    Description = "Windows Hello VBS hardware authenticator"
+                }
             };
             _entries.Add(new Guid(msftWhfbHardwareVbs.AaGuid), msftWhfbHardwareVbs);
 
@@ -149,12 +258,18 @@ namespace Fido2NetLib
             {
                 AaGuid = soloMetadataStatement.AaGuid,
                 Url = "https://raw.githubusercontent.com/solokeys/solo/master/metadata/Solo-FIDO2-CTAP2-Authenticator.json",
-                StatusReports = new StatusReport[] { new StatusReport() { Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED } },
+                StatusReports = new StatusReport[]
+                {
+                    new StatusReport
+                    {
+                        Status = AuthenticatorStatus.NOT_FIDO_CERTIFIED
+                    }
+                },
                 MetadataStatement = soloMetadataStatement
             };
             _entries.Add(new Guid(soloKeysSolo.AaGuid), soloKeysSolo);
 
-            foreach(var entry in _entries)
+            foreach (var entry in _entries)
             {
                 entry.Value.MetadataStatement.AaGuid = entry.Value.AaGuid;
             }
