@@ -7,8 +7,7 @@
         public static readonly AttestationType Self = new AttestationType("self");
         public static readonly AttestationType AttCa = new AttestationType("attca");
         public static readonly AttestationType ECDAA = new AttestationType("ecdaa");
-
-
+        
         private AttestationType(string value) : base(value)
         {
         }
